% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-boxLegendGrob.R
\name{boxLegendGrob}
\alias{boxLegendGrob}
\title{Create a filled-box legend}
\usage{
boxLegendGrob(labels, fills, name = NULL, vp = NULL, gp = NULL)
}
\arguments{
\item{labels}{Character, the legend labels}

\item{fills}{Colors to fill the slices}

\item{name, vp, gp}{Passed to \link[grid:grid.frame]{grid::frameGrob}}
}
\description{
Create a filled-box legend
}
\details{
Not yet exported. Need to break out the creation of viewports and grobs as
done in the exported grobs. This will allow better grobEdit methods, which
also needs to be created for the boxLegendGrob.
Also need to do some input checks.

Also, if \code{grid::legendGrob} gets updated to use the 'has.fill' option
this function should be removed and \code{grid::legendGrob} can be used
instead.
}
