% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/methods-TxpModel.R,
%   R/methods-TxpModelList.R, R/txpCalculateScores.R
\name{txpCalculateScores}
\alias{txpCalculateScores}
\alias{txpCalculateScores,TxpModel,data.frame-method}
\alias{txpCalculateScores,TxpModelList,data.frame-method}
\alias{txpCalculateScores,list,data.frame-method}
\title{Calculate ToxPi Scores for the given model and input data}
\usage{
txpCalculateScores(
  model,
  input,
  id.var = NULL,
  rank.ties.method = NULL,
  negative.value.handling = NULL
)

\S4method{txpCalculateScores}{TxpModel,data.frame}(
  model,
  input,
  id.var = NULL,
  rank.ties.method = NULL,
  negative.value.handling = NULL
)

\S4method{txpCalculateScores}{TxpModelList,data.frame}(
  model,
  input,
  id.var = NULL,
  rank.ties.method = NULL,
  negative.value.handling = NULL
)

\S4method{txpCalculateScores}{list,data.frame}(
  model,
  input,
  id.var = NULL,
  rank.ties.method = NULL,
  negative.value.handling = NULL
)
}
\arguments{
\item{model}{S4 \link{TxpModel} object or \link{TxpModelList} object. The txpModel object to be used in the calculation}

\item{input}{data.frame. The input dataframe to be analyzed that corresponds to the txpModel object provided}

\item{id.var}{Optional scalar character or numeric. The name or index of the column in the input containing
unique identifiers for the rows. If NULL, the row indices of the original data will be used as
unique names for identification}

\item{rank.ties.method}{Scalar character. Optionally overwrite \code{rankTies} in
the provided txpModel with a new method for handling rank ties. Options are
average', 'first', 'last', 'random', 'max', 'min'; when NULL keeps method stated in model}

\item{negative.value.handling}{Scalar character. Optionally overwrite \code{negativeHandling} in
the provided txpModel with a new method for handling negative data. Options are
'keep' and missing'; when NULL keeps method stated in model}
}
\value{
\link{TxpResult} or \link{TxpResultList} object
}
\description{
Calculate ToxPi Scores for the given model and input data
}
\details{
\code{txpCalculateScores} is implemented as an S4 generic function with methods
for \link{TxpModel} and \link{TxpModelList}.

Ranks are calculated such that the highest ToxPi score has a rank of 1.

Missingness is determined after applying input-level transformations but
before applying slice-level transformations.
}
\examples{
## Load example dataset & model; see ?TxpModel for building model objects
data(txp_example_input, package = "toxpiR")
data(txp_example_model, package = "toxpiR")
data(txp_example_input_CI, package = "toxpiR")
data(txp_example_model_CI, package = "toxpiR")
data(txp_example_results_CI, package = "toxpiR")
## Calculate scores for single model; returns TxpResult object
res <- txpCalculateScores(model = txp_example_model, 
                          input = txp_example_input, 
                          id.var = "name")

## Calculate scores for list of models; returns TxpResultList object
txpCalculateScores(model = TxpModelList(m1 = txp_example_model, 
                                        m2 = txp_example_model), 
                   input = txp_example_input, 
                   id.var = "name")
resLst <- txpCalculateScores(model = list(m1 = txp_example_model, 
                                          m2 = txp_example_model), 
                             input = txp_example_input, 
                             id.var = "name")

}
\seealso{
\link{TxpModel}, \link{TxpResult}, \link{TxpResultParam}
}
