% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txpExportCSV.R
\name{txpExportCSV}
\alias{txpExportCSV}
\title{Export comma-separated file containing txpModel and raw data}
\usage{
txpExportCSV(
  input,
  model,
  id.var = NULL,
  fills = NULL,
  format = "toxpiR",
  fileName = "txpModel.csv"
)
}
\arguments{
\item{input}{data.frame. The input dataframe to be analyzed that corresponds to the txpModel object provided}

\item{model}{S4 \link{TxpModel} object or \link{TxpModelList} object. The txpModel object to be used in the calculation}

\item{id.var}{Optional scalar character or numeric. The name or index of the column in the input containing
unique identifiers for the rows. If NULL, the row indices of the original data will be used as
unique names for identification}

\item{fills}{Colors to fill the slices}

\item{format}{Character scalar, export format for csv file. Options are
'toxpiR' and 'gui'. Defaults to 'toxpiR'.}

\item{fileName}{Character scalar, the path to the output file}
}
\description{
Export comma-separated file containing txpModel and raw data
}
\details{
This function exports data and models to csv files for easy sharing and
recreation. The suggested format is toxpiR, whereas gui is provided for
backwards compatibility with the toxpi GUI. The GUI format differs in several
meaningful ways: (1) the GUI only allows for integer weights; (2) the GUI can
only apply metric transformations; (3) all metrics within a slice undergo one
common metric transformation; (4) metric transformations can only be chosen
from a predefined subset; (5) negative handling can only be treated as
missing; (6) the GUI doesn't allow for user provided confidence intervals

\code{txpExportCSV} with \code{format = "gui"} will not work for models with
non-integer weights, negative handling set to missing, nor slices containing
upper/lower CIs

The GUI only applies a single transformation function to every input within
a slice, and only functions from a pre-determined list; \code{toxpiR} allows
users to apply any valid function individually to each input, then a second
transformation function on the summed slice values. Because of this
complexity, any exported models with slice-level transformation functions
will not export at the input level. In other words, the export will have only
the final slice scores. Otherwise, all input-level transformations will be
performed, and the export will contain transformed input-level data with
the \code{linear(x)} GUI transformation.
}
