% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{txpGenerics}
\alias{txpGenerics}
\alias{txpValueNames}
\alias{txpValueNames<-}
\alias{txpTransFuncs}
\alias{txpTransFuncs<-}
\alias{txpValueNamesLower}
\alias{txpValueNamesLower<-}
\alias{txpTransFuncsLower}
\alias{txpTransFuncsLower<-}
\alias{txpValueNamesUpper}
\alias{txpValueNamesUpper<-}
\alias{txpTransFuncsUpper}
\alias{txpTransFuncsUpper<-}
\alias{txpSlices}
\alias{txpSlices<-}
\alias{txpWeights}
\alias{txpWeights<-}
\alias{negativeHandling}
\alias{negativeHandling<-}
\alias{rankTies}
\alias{rankTies<-}
\alias{txpScores}
\alias{txpScoresLower}
\alias{txpScoresUpper}
\alias{txpSliceScores}
\alias{txpSliceScoresLower}
\alias{txpSliceScoresUpper}
\alias{txpModel}
\alias{txpIDs}
\alias{txpIDs<-}
\alias{txpRanks}
\alias{txpRanksLower}
\alias{txpRanksUpper}
\alias{txpMissing}
\alias{txpResultParam}
\title{toxpiR package generics}
\usage{
txpValueNames(x, ...)

txpValueNames(x, ...) <- value

txpTransFuncs(x, ...)

txpTransFuncs(x, ...) <- value

txpValueNamesLower(x, ...)

txpValueNamesLower(x, ...) <- value

txpTransFuncsLower(x, ...)

txpTransFuncsLower(x, ...) <- value

txpValueNamesUpper(x, ...)

txpValueNamesUpper(x, ...) <- value

txpTransFuncsUpper(x, ...)

txpTransFuncsUpper(x, ...) <- value

txpSlices(x, ...)

txpSlices(x, ...) <- value

txpWeights(x, ...)

txpWeights(x, ...) <- value

negativeHandling(x, ...)

negativeHandling(x, ...) <- value

rankTies(x, ...)

rankTies(x, ...) <- value

txpScores(x, ...)

txpScoresLower(x, ...)

txpScoresUpper(x, ...)

txpSliceScores(x, ...)

txpSliceScoresLower(x, ...)

txpSliceScoresUpper(x, ...)

txpModel(x, ...)

txpIDs(x, ...)

txpIDs(x, ...) <- value

txpRanks(x, ...)

txpRanksLower(x, ...)

txpRanksUpper(x, ...)

txpMissing(x, ...)

txpResultParam(x, ...)
}
\arguments{
\item{x}{toxpiR S4 object}

\item{...}{Included for extendability; not currently used}

\item{value}{Replacement value}
}
\value{
See specific methods for details.
}
\description{
toxpiR package generics; see class man pages for associated
methods
}
