% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txpImportCSV.R
\name{txpImportCSV}
\alias{txpImportCSV}
\title{Import data file generated by ToxPi Gui or toxpiR Package}
\usage{
txpImportCSV(csvDataFile)
}
\arguments{
\item{csvDataFile}{Character scalar, the path to a 'data' export from the
ToxPi GUI or toxpiR package}
}
\value{
\code{list} with \verb{$model} containing \link{TxpModel} object; \verb{$input}
containing \code{data.frame} with input data; \verb{$fills} containing a vector
of fill colors.
}
\description{
Import data file generated by ToxPi Gui or toxpiR Package
}
\details{
This function takes the '_data.csv' files generated by the GUI or R package.
See \url{https://toxpi.org} for more information.

Because of the way toxpiR implements transformation functions, there is not
a way currently to use the GUI 'hitcount' function.
}
