% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txpImportGui.R
\name{txpImportGui}
\alias{txpImportGui}
\title{Import data file generated by ToxPi GUI (Deprecated)}
\usage{
txpImportGui(guiDataFile)
}
\arguments{
\item{guiDataFile}{Character scalar, the path to a 'data' export from the
ToxPi GUI}
}
\value{
\code{list} with \verb{$model} containing \link{TxpModel} object; \verb{$input}
containing \code{data.frame} with input data; \verb{$fills} containing a vector
of fill colors.
}
\description{
This function is deprecated. Please use \code{\link[=txpImportCSV]{txpImportCSV()}} instead.
}
\details{
This function takes the '_data.csv' files generated by the GUI.
See \url{https://toxpi.github.io/} for more information.

Because of the way toxpiR implements transformation functions, there is not
a way currently to use the GUI 'hitcount' function.
}
\seealso{
\code{\link[=txpImportCSV]{txpImportCSV()}}
}
\keyword{internal}
