% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-txpModelBuilder.R
\name{txpModelBuilder}
\alias{txpModelBuilder}
\title{Shiny App for building ToxPi models}
\usage{
txpModelBuilder(dataFile, id = NULL)
}
\arguments{
\item{dataFile}{A data frame or a file path to a \code{.csv} file. If the input is
a file path, the file must be a csv containing data structured as a typical
CSV, where the first row is expected to be a header}
}
\value{
No return. Models can be exported from the application as csv or rds
files
}
\description{
Opens a Shiny App for building and exporting a ToxPi model built
from a specific dataset
}
\details{
This function takes a csv or r dataframe as input and opens a shiny
application for building and exporting a ToxPi model.

The input is expected to contain samples as the rows, and corresponding
features/metrics as columns. For a csv, the first row must contain the column
names. For an r dataframe, \code{colnames} must be set to the corresponding names
}
\examples{
library(toxpiR)
txpModelBuilder('data/size_small.csv')

}
