% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/methods-TxpResultParam.R
\docType{class}
\name{TxpResultParam-class}
\alias{TxpResultParam-class}
\alias{TxpResultParam}
\title{ToxPi Result Parameters}
\arguments{
\item{rank.ties.method}{Passed to \code{rank.ties.method} slot}

\item{negative.value.handling}{Passed to \code{negative.value.handling} slot}
}
\description{
S4 class to store ToxPi result calculation parameters
}
\section{Slots}{

\describe{
\item{\code{rank.ties.method}}{Character scalar, method used to calculate score
ranks passed to \link[base:rank]{base::rank}}

\item{\code{negative.value.handling}}{Character scalar, how negative values are
handled, see details}
}}

\seealso{
\link{txpCalculateScores}, \link{TxpResult}
}
