% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxpiR-package.R
\docType{package}
\name{toxpiR-package}
\alias{toxpiR}
\alias{toxpiR-package}
\title{toxpiR: Create ToxPi Prioritization Models}
\description{
Enables users to build 'ToxPi' prioritization models and provides functionality within the grid framework for plotting ToxPi graphs. 'toxpiR' allows for more customization than the 'ToxPi GUI' (<https://toxpi.org>) and integration into existing workflows for greater ease-of-use, reproducibility, and transparency. toxpiR package behaves nearly identically to the GUI; the package documentation includes notes about all differences. The vignettes download example files from <https://github.com/ToxPi/ToxPi-example-files>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ToxPi/toxpiR}
  \item \url{https://toxpi.github.io/toxpiR/}
  \item Report bugs at \url{https://github.com/ToxPi/toxpiR/issues}
}

}
\author{
\strong{Maintainer}: Dayne L Filer \email{dayne.filer@gmail.com} (\href{https://orcid.org/0000-0002-3443-5315}{ORCID}) [funder]

Authors:
\itemize{
  \item Dillon T Lloyd
  \item Preethi Thunga (\href{https://orcid.org/0000-0001-5447-0129}{ORCID})
  \item Skylar W Marvel
  \item David M Reif \email{reif.david@gmail.com} (\href{https://orcid.org/0000-0001-7815-6767}{ORCID}) [funder]
}

Other contributors:
\itemize{
  \item Alison A Motsinger-Reif (\href{https://orcid.org/0000-0003-1346-2493}{ORCID}) [funder]
}

}
\keyword{internal}
