% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{txpGenerics}
\alias{txpGenerics}
\alias{txpValueNames}
\alias{txpValueNames<-}
\alias{txpTransFuncs}
\alias{txpTransFuncs<-}
\alias{txpLowerNames}
\alias{txpLowerNames<-}
\alias{txpLowerFuncs}
\alias{txpLowerFuncs<-}
\alias{txpUpperNames}
\alias{txpUpperNames<-}
\alias{txpUpperFuncs}
\alias{txpUpperFuncs<-}
\alias{txpSlices}
\alias{txpSlices<-}
\alias{txpWeights}
\alias{txpWeights<-}
\alias{negativeHandling}
\alias{negativeHandling<-}
\alias{rankTies}
\alias{rankTies<-}
\alias{txpScores}
\alias{txpSliceScores}
\alias{txpSliceLows}
\alias{txpSliceUps}
\alias{txpModel}
\alias{txpIDs}
\alias{txpIDs<-}
\alias{txpRanks}
\alias{txpMissing}
\alias{txpResultParam}
\title{toxpiR package generics}
\usage{
txpValueNames(x, ...)

txpValueNames(x, ...) <- value

txpTransFuncs(x, ...)

txpTransFuncs(x, ...) <- value

txpLowerNames(x, ...)

txpLowerNames(x, ...) <- value

txpLowerFuncs(x, ...)

txpLowerFuncs(x, ...) <- value

txpUpperNames(x, ...)

txpUpperNames(x, ...) <- value

txpUpperFuncs(x, ...)

txpUpperFuncs(x, ...) <- value

txpSlices(x, ...)

txpSlices(x, ...) <- value

txpWeights(x, ...)

txpWeights(x, ...) <- value

negativeHandling(x, ...)

negativeHandling(x, ...) <- value

rankTies(x, ...)

rankTies(x, ...) <- value

txpScores(x, ...)

txpSliceScores(x, ...)

txpSliceLows(x, ...)

txpSliceUps(x, ...)

txpModel(x, ...)

txpIDs(x, ...)

txpIDs(x, ...) <- value

txpRanks(x, ...)

txpMissing(x, ...)

txpResultParam(x, ...)
}
\arguments{
\item{x}{toxpiR S4 object}

\item{...}{Included for extendability; not currently used}

\item{value}{Replacement value}
}
\value{
See specific methods for details.
}
\description{
toxpiR package generics; see class man pages for associated
methods
}
