% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{txpGenerics}
\alias{txpGenerics}
\alias{txpValueNames}
\alias{txpValueNames<-}
\alias{txpTransFuncs}
\alias{txpTransFuncs<-}
\alias{txpSlices}
\alias{txpSlices<-}
\alias{txpWeights}
\alias{txpWeights<-}
\alias{txpScores}
\alias{txpSliceScores}
\alias{txpModel}
\alias{txpIDs}
\alias{txpIDs<-}
\alias{txpRanks}
\title{toxpiR package generics}
\usage{
txpValueNames(x, ...)

txpValueNames(x, ...) <- value

txpTransFuncs(x, ...)

txpTransFuncs(x, ...) <- value

txpSlices(x, ...)

txpSlices(x, ...) <- value

txpWeights(x, ...)

txpWeights(x, ...) <- value

txpScores(x, ...)

txpSliceScores(x, ...)

txpModel(x, ...)

txpIDs(x, ...)

txpIDs(x, ...) <- value

txpRanks(x, ...)
}
\arguments{
\item{x}{toxpiR S4 object}

\item{...}{Included for extendability; not currently used}

\item{value}{Replacement value}
}
\value{
See specific methods for details.
}
\description{
toxpiR package generics; see class man pages for associated
methods
}
